Split base runner, git runner, and non-runner execution
The non-runner execution (BatchedCommand) is put back in its
original place, where it was before this RF.
